Create a Dirichlet-tree for modelling ranked ballots

A dirichlet_tree object represents a Dirichlet-tree distribution on ballots. By specifying the tree structure for the ranked ballots, the Dirichlet-tree is initialized with the same prior structure described by Everest et al. (2022) . There are methods provided for observing data (to obtain a posterior distribution) along with methods to sample election outcomes and sets of ballots from the posterior predictive distribution.

Format

An R6Class generator object.

References

Everest F, Blom M, Stark PB, Stuckey PJ, Teague V, Vukcevic D (2022). “Ballot-Polling Audits of Instant-Runoff Voting Elections with a Dirichlet-Tree Model.” doi:10.48550/ARXIV.2209.03881 . .

Everest F, Blom M, Stark PB, Stuckey PJ, Teague V, Vukcevic D (2022). “Auditing Ranked Voting Elections with Dirichlet-Tree Models: First Steps.” doi:10.48550/ARXIV.2206.14605 . .

Active bindings

a0

Gets or sets the a0 parameter for the Dirichlet-tree.

min_depth

Gets or sets the min_depth parameter for the Dirichlet-tree.

max_depth

Gets or sets the max_depth parameter for the Dirichlet-tree.

vd

Gets or sets the vd parameter for the Dirichlet-tree.

Methods

Public methods

dirichlet_tree$new()

dirichlet_tree$print()

dirichlet_tree$update()

dirichlet_tree$reset()

dirichlet_tree$sample_posterior()

dirichlet_tree$sample_predictive()

Method new()

Create a new dirichlet_tree prior distribution with the specified tree structure. See Everest et al. (2022) for further details.

Usage

dirichlet_tree$new(
  candidates,
  min_depth = 0,
  max_depth = length(candidates) - 1,
  a0 = 1,
  vd = FALSE
)

Arguments

candidates

A character vector, with each element (must be unique) representing a single candidate.

min_depth

The minimum number of candidates which must be specified for a valid ballot in the election.

max_depth

The maximum number of candidates which can be specified for a valid ballot in the election.

a0

The prior parameter for the distribution.

vd

A flag which, when TRUE, employs a parameter structure which reduces to a regular Dirichlet distribution as described by Everest et al. (2022) .

Returns

A new dirichlet_tree prior.

Examples

dtree <- dirichlet_tree$new(candidates = LETTERS, a0 = 1., min_depth = 1)

Method print()

print shows some details of the distribution and its parameters.

Usage

dirichlet_tree$print()

Returns

The dirichlet_tree object.

Method update()

Updates the dirichlet_tree object with observations of ballots. This updates the parameter structure of the tree to yield the posterior Dirichlet-tree, as described in Everest et al. (2022) .

Usage

dirichlet_tree$update(ballots)

Arguments

ballots

A set of ballots to observe - must be of class ranked_ballots.

Returns

The dirichlet_tree object.

Examples

dirichlet_tree$new(
  candidates = LETTERS
)$update(
  ranked_ballots(c("A", "B", "C"))
)

Method reset()

Resets the dirichlet_tree observations to revert the parameter structure back to the originally specified prior.

Usage

dirichlet_tree$reset()

Returns

The dirichlet_tree object.

Examples

dirichlet_tree$new(
  candidates = LETTERS
)$update(
  ranked_ballots(c("A", "B", "C"))
)$reset()

Method sample_posterior()

Draws sets of ballots from independent realizations of the Dirichlet-tree posterior, then determines the probability for each candidate being elected by aggregating the results of the social choice function. See Everest et al. (2022) for details.

Usage

dirichlet_tree$sample_posterior(n_elections, n_ballots, n_winners = 1)

Arguments

n_elections

An integer representing the number of elections to generate. A higher number yields higher precision in the output probabilities.

n_ballots

An integer representing the total number of ballots cast in the election.

n_winners

The number of candidates elected in each election.

Returns

A numeric vector containing the probabilities for each candidate being elected.

Examples

dirichlet_tree$new(
  candidates = LETTERS,
  a0 = 1.,
  min_depth = 3,
  max_depth = 6,
  vd = FALSE
)$update(
  ranked_ballots(c("A","B","C"))
)$sample_posterior(
  n_elections = 100,
  n_ballots = 1000
)

Method sample_predictive()

sample_predictive draws ballots from a multinomial distribution with ballot probabilities obtained from a single realization of the Dirichlet-tree posterior on the ranked ballots. See Everest et al. (2022) for details.

Usage

dirichlet_tree$sample_predictive(n_ballots)

Arguments

n_ballots

An integer representing the total number of ballots cast in the election.

Returns

A ranked_ballots object containing n_ballots ballots drawn from a single realisation of the posterior Dirichlet-tree.

Examples

dirichlet_tree$new(
  candidates = LETTERS,
  a0 = 1.,
  min_depth = 3,
  max_depth = 6,
  vd = FALSE
)$update(
  ranked_ballots(c("A","B","C"))
)$sample_predictive(
  n_ballots = 10
)

Examples

## ------------------------------------------------
## Method `dirichlet_tree$new`
## ------------------------------------------------

dtree <- dirichlet_tree$new(candidates = LETTERS, a0 = 1., min_depth = 1)

## ------------------------------------------------
## Method `dirichlet_tree$update`
## ------------------------------------------------

dirichlet_tree$new(
  candidates = LETTERS
)$update(
  ranked_ballots(c("A", "B", "C"))
)

## ------------------------------------------------
## Method `dirichlet_tree$reset`
## ------------------------------------------------

dirichlet_tree$new(
  candidates = LETTERS
)$update(
  ranked_ballots(c("A", "B", "C"))
)$reset()

## ------------------------------------------------
## Method `dirichlet_tree$sample_posterior`
## ------------------------------------------------

dirichlet_tree$new(
  candidates = LETTERS,
  a0 = 1.,
  min_depth = 3,
  max_depth = 6,
  vd = FALSE
)$update(
  ranked_ballots(c("A","B","C"))
)$sample_posterior(
  n_elections = 100,
  n_ballots = 1000
)
#>    A    B    C    D    E    F    G    H    I    J    K    L    M    N    O    P 
#> 0.14 0.02 0.05 0.03 0.04 0.04 0.02 0.05 0.05 0.03 0.05 0.02 0.03 0.03 0.02 0.05 
#>    Q    R    S    T    U    V    W    X    Y    Z 
#> 0.07 0.05 0.03 0.01 0.03 0.03 0.03 0.02 0.04 0.02 

## ------------------------------------------------
## Method `dirichlet_tree$sample_predictive`
## ------------------------------------------------

dirichlet_tree$new(
  candidates = LETTERS,
  a0 = 1.,
  min_depth = 3,
  max_depth = 6,
  vd = FALSE
)$update(
  ranked_ballots(c("A","B","C"))
)$sample_predictive(
  n_ballots = 10
)
#> [[1]]
#> [1] "A" "K" "D" "E" "S" "Y"
#> 
#> [[2]]
#> [1] "E" "M" "H" "U"
#> 
#> [[3]]
#> [1] "H" "D" "J" "G" "C" "A"
#> 
#> [[4]]
#> [1] "H" "T" "U" "D" "I" "E"
#> 
#> [[5]]
#> [1] "I" "W" "B" "Q" "Y" "F"
#> 
#> [[6]]
#> [1] "J" "M" "Z" "N" "W" "E"
#> 
#> [[7]]
#> [1] "T" "J" "R" "B" "W" "X"
#> 
#> [[8]]
#> [1] "Y" "N" "W" "O" "V" "Z"
#> 
#> [[9]]
#> [1] "Z" "K" "I" "B" "L" "T"
#> 
#> [[10]]
#> [1] "Z" "N" "R" "M" "O" "X"
#> 
#> attr(,"class")
#> [1] "ranked_ballots"
#> attr(,"candidates")
#>  [1] "Z" "X" "Y" "V" "T" "Q" "O" "N" "A" "C" "W" "U" "D" "I" "B" "H" "S" "E" "F"
#> [20] "G" "K" "J" "P" "L" "R" "M"